Refactor TensorAt   locations* must be const and int64_t since our dims are int64_t   Remove unnecessary copy of location.   Remove unnecesary casting and C-casting. Simplify implementation.

+++ include/onnxruntime/core/session/onnxruntime_cxx_inline.h
@@ -771,9 +771,9 @@ const T* Value::GetTensorData() const {
 }
 
 template <typename T>
-inline T Value::At(const std::initializer_list<size_t>& location) {
+inline T Value::At(const std::initializer_list<int64_t>& location) {
   T* out;
-  std::vector<size_t> location_ = location;
+  std::vector<int64_t> location_ = location;
   ThrowOnError(GetApi().TensorAt(p_, location_.data(), location_.size(), (void**)&out));
   return *out;
 }
